Extract DummyDataUtilities
