+++ xcode/Subconscious/Shared/Models/StoryUser.swift
@@ -29,45 +29,3 @@ struct StoryUser:
         """
     }
 }
-
-protocol Generatable {
-    static func generateRandomTestInstance() -> Self
-}
-
-extension Bool: Generatable {
-    static func generateRandomTestInstance() -> Bool {
-        random()
-    }
-}
-
-extension Did: Generatable {
-    static func generateRandomTestInstance() -> Did {
-        let letters = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
-        let randomString = String((0..<32).map{ _ in letters.randomElement()! })
-        return Did(did: "did:key:\(randomString)")! // OK to do this for test data
-    }
-}
-
-extension Petname: Generatable {
-    static func generateRandomTestInstance() -> Petname {
-        let letters = "abcdefghijklmnopqrstuvwxyz-_0123456789"
-        let randomString = String((0..<8).map{ _ in letters.randomElement()! })
-        return Petname(randomString)! // OK to do this for test data
-    }
-}
-
-extension StoryUser: Generatable {
-    static func generateRandomTestInstance() -> StoryUser {
-        StoryUser(
-            user: UserProfile(
-                did: Did.generateRandomTestInstance(),
-                petname: Petname.generateRandomTestInstance(),
-                pfp: "pfp-dog",
-                bio: "Ploofy snooflewhumps burbled, outflonking the zibber-zabber.",
-                category: [UserCategory.human, UserCategory.geist].randomElement()!
-            ),
-            following: Bool.generateRandomTestInstance()
-        )
-    }
-}
-
